+++ include/MqttHandleVedirect.h
@@ -21,7 +21,7 @@ public:
     void forceUpdate();
 private:
     void loop();
-    std::map<std::string, VeDirectMpptController::veMpptStruct> _kvFrames;
+    std::map<std::string, VeDirectMpptController::data_t> _kvFrames;
 
     Task _loopTask;
 
@@ -34,7 +34,7 @@ private:
     bool _PublishFull;
 
     void publish_mppt_data(const VeDirectMpptController::spData_t &spMpptData,
-                           VeDirectMpptController::veMpptStruct &frame) const;
+                           const VeDirectMpptController::data_t &frame) const;
 };
 
 extern MqttHandleVedirectClass MqttHandleVedirect;
