+++ lib/VeDirectFrameHandler/VeDirectFrameHandler.cpp
@@ -58,7 +58,8 @@ class Silent : public Print {
 
 static Silent MessageOutputDummy;
 
-VeDirectFrameHandler::VeDirectFrameHandler() :
+template<typename T>
+VeDirectFrameHandler<T>::VeDirectFrameHandler() :
 	_msgOut(&MessageOutputDummy),
 	_lastUpdate(0),
 	_state(IDLE),
@@ -72,7 +73,8 @@ VeDirectFrameHandler::VeDirectFrameHandler() :
 {
 }
 
-void VeDirectFrameHandler::init(char const* who, int8_t rx, int8_t tx, Print* msgOut, bool verboseLogging, uint16_t hwSerialPort)
+template<typename T>
+void VeDirectFrameHandler<T>::init(char const* who, int8_t rx, int8_t tx, Print* msgOut, bool verboseLogging, uint16_t hwSerialPort)
 {
 	_vedirectSerial = std::make_unique<HardwareSerial>(hwSerialPort);
 	_vedirectSerial->begin(19200, SERIAL_8N1, rx, tx);
@@ -84,7 +86,8 @@ void VeDirectFrameHandler::init(char const* who, int8_t rx, int8_t tx, Print* ms
 	if (_verboseLogging) { _msgOut->printf("%s init complete\r\n", _logId); }
 }
 
-void VeDirectFrameHandler::dumpDebugBuffer() {
+template<typename T>
+void VeDirectFrameHandler<T>::dumpDebugBuffer() {
 	_msgOut->printf("%s serial input (%d Bytes):", _logId, _debugIn);
 	for (int i = 0; i < _debugIn; ++i) {
 		if (i % 16 == 0) {
@@ -96,7 +99,16 @@ void VeDirectFrameHandler::dumpDebugBuffer() {
 	_debugIn = 0;
 }
 
-void VeDirectFrameHandler::loop()
+template<typename T>
+void VeDirectFrameHandler<T>::reset()
+{
+	_checksum = 0;
+	_state = IDLE;
+	_textData.clear();
+}
+
+template<typename T>
+void VeDirectFrameHandler<T>::loop()
 {
 	while ( _vedirectSerial->available()) {
 		rxData(_vedirectSerial->read());
@@ -109,8 +121,7 @@ void VeDirectFrameHandler::loop()
 	if (IDLE != _state && _lastByteMillis + 500 < millis()) {
 		_msgOut->printf("%s Resetting state machine (was %d) after timeout\r\n", _logId, _state);
 		if (_verboseLogging) { dumpDebugBuffer(); }
-		_checksum = 0;
-		_state = IDLE;
+		reset();
 	}
 }
 
@@ -119,7 +130,8 @@ void VeDirectFrameHandler::loop()
  *  This function is called by loop() which passes a byte of serial data
  *  Based on Victron's example code. But using String and Map instead of pointer and arrays
  */
-void VeDirectFrameHandler::rxData(uint8_t inbyte)
+template<typename T>
+void VeDirectFrameHandler<T>::rxData(uint8_t inbyte)
 {
 	if (_verboseLogging) {
 		_debugBuffer[_debugIn] = inbyte;
@@ -186,7 +198,7 @@ void VeDirectFrameHandler::rxData(uint8_t inbyte)
 		case '\n':
 			if ( _textPointer < (_value + sizeof(_value)) ) {
 				*_textPointer = 0; // make zero ended
-				textRxEvent(_name, _value);
+				_textData.push_back({_name, _value});
 			}
 			_state = RECORD_BEGIN;
 			break;
@@ -201,14 +213,18 @@ void VeDirectFrameHandler::rxData(uint8_t inbyte)
 		break;
 	case CHECKSUM:
 	{
-		bool valid = _checksum == 0;
-		if (!valid) {
-			_msgOut->printf("%s checksum 0x%02x != 0, invalid frame\r\n", _logId, _checksum);
-		}
 		if (_verboseLogging) { dumpDebugBuffer(); }
-		_checksum = 0;
-		_state = IDLE;
-		if (valid) { frameValidEvent(); }
+		if (_checksum == 0) {
+			for (auto const& event : _textData) {
+				processTextData(event.first, event.second);
+			}
+			_lastUpdate = millis();
+			frameValidEvent();
+		}
+		else {
+			_msgOut->printf("%s checksum 0x%02x != 0x00, invalid frame\r\n", _logId, _checksum);
+		}
+		reset();
 		break;
 	}
 	case RECORD_HEX:
@@ -218,41 +234,44 @@ void VeDirectFrameHandler::rxData(uint8_t inbyte)
 }
 
 /*
- * textRxEvent
  * This function is called every time a new name/value is successfully parsed.  It writes the values to the temporary buffer.
  */
-bool VeDirectFrameHandler::textRxEvent(char* name, char* value, veStruct& frame) {
+template<typename T>
+void VeDirectFrameHandler<T>::processTextData(std::string const& name, std::string const& value) {
 	if (_verboseLogging) {
-		_msgOut->printf("%s Text Event %s: Value: %s\r\n",
-				_logId, name, value );
+		_msgOut->printf("%s Text Data '%s' = '%s'\r\n",
+				_logId, name.c_str(), value.c_str());
 	}
 
-	if (strcmp(name, "PID") == 0) {
-		frame.PID = strtol(value, nullptr, 0);
-		return true;
+	if (processTextDataDerived(name, value)) { return; }
+
+	if (name == "PID") {
+		_tmpFrame.PID = strtol(value.c_str(), nullptr, 0);
+		return;
 	}
 
-	if (strcmp(name, "SER") == 0) {
-		strcpy(frame.SER, value);
-		return true;
+	if (name == "SER") {
+		strcpy(_tmpFrame.SER, value.c_str());
+		return;
 	}
 
-	if (strcmp(name, "FW") == 0) {
-		strcpy(frame.FW, value);
-		return true;
+	if (name == "FW") {
+		strcpy(_tmpFrame.FW, value.c_str());
+		return;
 	}
 
-	if (strcmp(name, "V") == 0) {
-		frame.V = round(atof(value) / 10.0) / 100.0;
-		return true;
+	if (name == "V") {
+		_tmpFrame.V = round(atof(value.c_str()) / 10.0) / 100.0;
+		return;
 	}
 
-	if (strcmp(name, "I") == 0) {
-		frame.I = round(atof(value) / 10.0) / 100.0;
-		return true;
+	if (name == "I") {
+		_tmpFrame.I = round(atof(value.c_str()) / 10.0) / 100.0;
+		return;
 	}
 
-	return false;
+	_msgOut->printf("%s Unknown text data '%s' (value '%s')\r\n",
+			_logId, name.c_str(), value.c_str());
 }
 
 
@@ -261,7 +280,9 @@ bool VeDirectFrameHandler::textRxEvent(char* name, char* value, veStruct& frame)
  *  hexRxEvent
  *  This function records hex answers or async messages
  */
-int VeDirectFrameHandler::hexRxEvent(uint8_t inbyte) {
+template<typename T>
+int VeDirectFrameHandler<T>::hexRxEvent(uint8_t inbyte)
+{
 	int ret=RECORD_HEX; // default - continue recording until end of frame
 
 	switch (inbyte) {
@@ -282,138 +303,14 @@ int VeDirectFrameHandler::hexRxEvent(uint8_t inbyte) {
 	return ret;
 }
 
-bool VeDirectFrameHandler::isDataValid(veStruct const& frame) const {
-	return strlen(frame.SER) > 0 && _lastUpdate > 0 && (millis() - _lastUpdate) < (10 * 1000);
+template<typename T>
+bool VeDirectFrameHandler<T>::isDataValid() const
+{
+	return strlen(_tmpFrame.SER) > 0 && _lastUpdate > 0 && (millis() - _lastUpdate) < (10 * 1000);
 }
 
-uint32_t VeDirectFrameHandler::getLastUpdate() const
+template<typename T>
+uint32_t VeDirectFrameHandler<T>::getLastUpdate() const
 {
 	return _lastUpdate;
-}
-
-/*
- * getPidAsString
- * This function returns the product id (PID) as readable text.
- */
-frozen::string const& VeDirectFrameHandler::veStruct::getPidAsString() const
-{
-	/**
-	 * this map is rendered from [1], which is more recent than [2]. Phoenix
-	 * inverters are not included in the map. unfortunately, the documents do
-	 * not fully align. PID 0xA07F is only present in [1]. PIDs 0xA048, 0xA110,
-	 * and 0xA111 are only present in [2]. PIDs 0xA06D and 0xA078 are rev3 in
-	 * [1] but rev2 in [2].
-	 *
-	 * [1] https://www.victronenergy.com/upload/documents/VE.Direct-Protocol-3.33.pdf
-	 * [2] https://www.victronenergy.com/upload/documents/BlueSolar-HEX-protocol.pdf
-	 */
-	static constexpr frozen::map<uint16_t, frozen::string, 105> values = {
-		{ 0x0203, "BMV-700" },
-		{ 0x0204, "BMV-702" },
-		{ 0x0205, "BMV-700H" },
-		{ 0x0300, "BlueSolar MPPT 70|15" },
-		{ 0xA040, "BlueSolar MPPT 75|50" },
-		{ 0xA041, "BlueSolar MPPT 150|35" },
-		{ 0xA042, "BlueSolar MPPT 75|15" },
-		{ 0xA043, "BlueSolar MPPT 100|15" },
-		{ 0xA044, "BlueSolar MPPT 100|30" },
-		{ 0xA045, "BlueSolar MPPT 100|50" },
-		{ 0xA046, "BlueSolar MPPT 150|70" },
-		{ 0xA047, "BlueSolar MPPT 150|100" },
-		{ 0xA048, "BlueSolar MPPT 75|50 rev2" },
-		{ 0xA049, "BlueSolar MPPT 100|50 rev2" },
-		{ 0xA04A, "BlueSolar MPPT 100|30 rev2" },
-		{ 0xA04B, "BlueSolar MPPT 150|35 rev2" },
-		{ 0xA04C, "BlueSolar MPPT 75|10" },
-		{ 0xA04D, "BlueSolar MPPT 150|45" },
-		{ 0xA04E, "BlueSolar MPPT 150|60" },
-		{ 0xA04F, "BlueSolar MPPT 150|85" },
-		{ 0xA050, "SmartSolar MPPT 250|100" },
-		{ 0xA051, "SmartSolar MPPT 150|100" },
-		{ 0xA052, "SmartSolar MPPT 150|85" },
-		{ 0xA053, "SmartSolar MPPT 75|15" },
-		{ 0xA054, "SmartSolar MPPT 75|10" },
-		{ 0xA055, "SmartSolar MPPT 100|15" },
-		{ 0xA056, "SmartSolar MPPT 100|30" },
-		{ 0xA057, "SmartSolar MPPT 100|50" },
-		{ 0xA058, "SmartSolar MPPT 150|35" },
-		{ 0xA059, "SmartSolar MPPT 150|100 rev2" },
-		{ 0xA05A, "SmartSolar MPPT 150|85 rev2" },
-		{ 0xA05B, "SmartSolar MPPT 250|70" },
-		{ 0xA05C, "SmartSolar MPPT 250|85" },
-		{ 0xA05D, "SmartSolar MPPT 250|60" },
-		{ 0xA05E, "SmartSolar MPPT 250|45" },
-		{ 0xA05F, "SmartSolar MPPT 100|20" },
-		{ 0xA060, "SmartSolar MPPT 100|20 48V" },
-		{ 0xA061, "SmartSolar MPPT 150|45" },
-		{ 0xA062, "SmartSolar MPPT 150|60" },
-		{ 0xA063, "SmartSolar MPPT 150|70" },
-		{ 0xA064, "SmartSolar MPPT 250|85 rev2" },
-		{ 0xA065, "SmartSolar MPPT 250|100 rev2" },
-		{ 0xA066, "BlueSolar MPPT 100|20" },
-		{ 0xA067, "BlueSolar MPPT 100|20 48V" },
-		{ 0xA068, "SmartSolar MPPT 250|60 rev2" },
-		{ 0xA069, "SmartSolar MPPT 250|70 rev2" },
-		{ 0xA06A, "SmartSolar MPPT 150|45 rev2" },
-		{ 0xA06B, "SmartSolar MPPT 150|60 rev2" },
-		{ 0xA06C, "SmartSolar MPPT 150|70 rev2" },
-		{ 0xA06D, "SmartSolar MPPT 150|85 rev3" },
-		{ 0xA06E, "SmartSolar MPPT 150|100 rev3" },
-		{ 0xA06F, "BlueSolar MPPT 150|45 rev2" },
-		{ 0xA070, "BlueSolar MPPT 150|60 rev2" },
-		{ 0xA071, "BlueSolar MPPT 150|70 rev2" },
-		{ 0xA072, "BlueSolar MPPT 150|45 rev3" },
-		{ 0xA073, "SmartSolar MPPT 150|45 rev3" },
-		{ 0xA074, "SmartSolar MPPT 75|10 rev2" },
-		{ 0xA075, "SmartSolar MPPT 75|15 rev2" },
-		{ 0xA076, "BlueSolar MPPT 100|30 rev3" },
-		{ 0xA077, "BlueSolar MPPT 100|50 rev3" },
-		{ 0xA078, "BlueSolar MPPT 150|35 rev3" },
-		{ 0xA079, "BlueSolar MPPT 75|10 rev2" },
-		{ 0xA07A, "BlueSolar MPPT 75|15 rev2" },
-		{ 0xA07B, "BlueSolar MPPT 100|15 rev2" },
-		{ 0xA07C, "BlueSolar MPPT 75|10 rev3" },
-		{ 0xA07D, "BlueSolar MPPT 75|15 rev3" },
-		{ 0xA07E, "SmartSolar MPPT 100|30 12V" },
-		{ 0xA07F, "All-In-1 SmartSolar MPPT 75|15 12V" },
-		{ 0xA102, "SmartSolar MPPT VE.Can 150|70" },
-		{ 0xA103, "SmartSolar MPPT VE.Can 150|45" },
-		{ 0xA104, "SmartSolar MPPT VE.Can 150|60" },
-		{ 0xA105, "SmartSolar MPPT VE.Can 150|85" },
-		{ 0xA106, "SmartSolar MPPT VE.Can 150|100" },
-		{ 0xA107, "SmartSolar MPPT VE.Can 250|45" },
-		{ 0xA108, "SmartSolar MPPT VE.Can 250|60" },
-		{ 0xA109, "SmartSolar MPPT VE.Can 250|70" },
-		{ 0xA10A, "SmartSolar MPPT VE.Can 250|85" },
-		{ 0xA10B, "SmartSolar MPPT VE.Can 250|100" },
-		{ 0xA10C, "SmartSolar MPPT VE.Can 150|70 rev2" },
-		{ 0xA10D, "SmartSolar MPPT VE.Can 150|85 rev2" },
-		{ 0xA10E, "SmartSolar MPPT VE.Can 150|100 rev2" },
-		{ 0xA10F, "BlueSolar MPPT VE.Can 150|100" },
-		{ 0xA110, "SmartSolar MPPT RS 450|100" },
-		{ 0xA111, "SmartSolar MPPT RS 450|200" },
-		{ 0xA112, "BlueSolar MPPT VE.Can 250|70" },
-		{ 0xA113, "BlueSolar MPPT VE.Can 250|100" },
-		{ 0xA114, "SmartSolar MPPT VE.Can 250|70 rev2" },
-		{ 0xA115, "SmartSolar MPPT VE.Can 250|100 rev2" },
-		{ 0xA116, "SmartSolar MPPT VE.Can 250|85 rev2" },
-		{ 0xA117, "BlueSolar MPPT VE.Can 150|100 rev2" },
-		{ 0xA340, "Phoenix Smart IP43 Charger 12|50 (1+1)" },
-		{ 0xA341, "Phoenix Smart IP43 Charger 12|50 (3)" },
-		{ 0xA342, "Phoenix Smart IP43 Charger 24|25 (1+1)" },
-		{ 0xA343, "Phoenix Smart IP43 Charger 24|25 (3)" },
-		{ 0xA344, "Phoenix Smart IP43 Charger 12|30 (1+1)" },
-		{ 0xA345, "Phoenix Smart IP43 Charger 12|30 (3)" },
-		{ 0xA346, "Phoenix Smart IP43 Charger 24|16 (1+1)" },
-		{ 0xA347, "Phoenix Smart IP43 Charger 24|16 (3)" },
-		{ 0xA381, "BMV-712 Smart" },
-		{ 0xA382, "BMV-710H Smart" },
-		{ 0xA383, "BMV-712 Smart Rev2" },
-		{ 0xA389, "SmartShunt 500A/50mV" },
-		{ 0xA38A, "SmartShunt 1000A/50mV" },
-		{ 0xA38B, "SmartShunt 2000A/50mV" },
-		{ 0xA3F0, "Smart BuckBoost 12V/12V-50A" },
-	};
-
-	return getAsString(values, PID);
 }
